Backtesting services, new charting Interface and the controller work except for the test graph implementation

+++ src/main/java/co/syngleton/chartomancer/api_requesting/BacktestingQueryService.java
@@ -7,5 +7,5 @@ import java.time.LocalDate;
 
 public interface BacktestingQueryService {
 
-    BacktestingResultsDTO getTradingSimulation(Symbol symbol, Timeframe timeframe, LocalDate startDate, LocalDate endDate, float accountBalance);
+    BacktestingResultsDTO getTradingSimulation(Symbol symbol, Timeframe timeframe, LocalDate startDate, LocalDate endDate, double accountBalance);
 }
